Rename TabletView folder
