+++ front/src/pages/TabletView/components/Notifications.tsx
