+++ front/src/pages/TabletView/components/OpenHoursSection.tsx
