+++ front/src/pages/TabletView/components/RangeOfficerStatusSection.tsx
