+++ front/src/pages/TabletView/components/TrackStatistics/TrackStatistics.test.tsx
