+++ front/src/pages/TabletView/components/TrackStatistics/TrackStatistics.tsx
