+++ front/src/pages/TabletView/rangeofficer.module.scss
